+++ dbms/src/Encryption/CompressedReadBufferFromFileProvider.cpp
@@ -12,9 +12,9 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include <Encryption/ChecksumReadBufferBuilder.h>
 #include <Encryption/CompressedReadBufferFromFileProvider.h>
 #include <Encryption/ReadBufferFromRandomAccessFileBuilder.h>
+#include <IO/Checksum/ChecksumReadBufferBuilder.h>
 #include <IO/WriteHelpers.h>
