+++ dbms/src/IO/Checksum/ChecksumBuffer.cpp
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include <Encryption/ChecksumBuffer.h>
+#include <IO/Checksum/ChecksumBuffer.h>
 
 namespace DB
 {
