+++ dbms/src/IO/Checksum/ChecksumBuffer.h
